# pages/api/profile.ts
import { NextApiRequest, NextApiResponse } from 'next';
import { unstable_getServerSession } from 'next-auth/next';
import authOptions from './auth/[...nextauth]';
import { getUser, updateProfile } from '@/controllers'
import { Session } from 'next-auth';
import { connect } from '@/database'
import { Profile } from '@/types/users';

/**
 * API handler to manage user profiles.
 * Supports GET for retrieving the user profile and POST for updating it.
 * The function first checks the user session for authentication.
 * Depending on the request method, it either retrieves or updates the user's profile.
 * 
 * @param {NextApiRequest} req - The API request object.
 * @param {NextApiResponse<Profile | String>} res - The API response object used to return the user profile or an error message.
 * @returns {Promise<void>} - A promise that resolves when the response is sent.
 */
export default async function handler(req: NextApiRequest, res: NextApiResponse<Profile | String>) {
    const session: Session = (await unstable_getServerSession(req, res, authOptions))!;
    if (!session) return res.status(401).send('Unauthorized');

    await connect();

    const { method } = req;
    switch (method) {
        case 'GET': {
            const profile = await getUser(session.user);
            return res.status(200).json(profile);
        }
        case 'POST': {
            const profile = await updateProfile(session.user, JSON.parse(req.body));
            return res.status(200).json(profile);
        }
        default:
            return res.status(405).send('Method Not Allowed');
    }
}