# swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Perfect Match API Documentation",
    "version": "1.0.0"
  },
  "components": {
    "schemas": {
      "Profile": {
        "type": "object",
        "properties": {
          "complete": {
            "type": "boolean"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "gender": {
            "type": "string"
          },
          "genderPref": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "age": {
            "type": "integer"
          },
          "height": {
            "type": "integer"
          },
          "city": {
            "type": "string"
          },
          "race": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "year": {
            "type": "string"
          },
          "college": {
            "type": "string"
          },
          "major": {
            "type": "string"
          },
          "commitment": {
            "type": "string"
          },
          "relationshipType": {
            "type": "string"
          },
          "agePref": {
            "type": "object",
            "properties": {
              "youngest": {
                "type": "integer"
              },
              "oldest": {
                "type": "integer"
              }
            }
          },
          "activities": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "describeYourself": {
            "type": "string"
          },
          "bio": {
            "type": "string"
          }
        }
      },
...
      "MatchFeedback": {
        "type": "object",
        "properties": {
          "overallRating": {
            "type": "number"
          },
          "topReasonForRating": {
            "type": "string"
          },
          "metMatch": {
            "type": "boolean"
          },
          "numberOfDates": {
            "type": "number"
          },
          "inRelationshipWithMatch": {
            "type": "boolean"
          },
          "additionalComments": {
            "type": "string"
          },
          "dateSubmitted": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  },
  "paths": {
    "/api/review/{matchid}": {
      "post": {
        "summary": "Update match review",
        "description": "Updates match review details for a given match ID.",
        "operationId": "updateMatchReview",
        "parameters": [
          {
            "name": "matchid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MatchFeedback"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Match review updated successfully"
          },
          "400": {
            "description": "Bad request, invalid inputs"
          }
        }
      }
    },
...
    "/api/profile": {
      "get": {
        "summary": "Retrieve user profile",
        "description": "Returns the user's profile details after authentication.",
        "responses": {
          "200": {
            "description": "User profile data retrieved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Profile"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized access."
          }
        }
      },
      "post": {
        "summary": "Update user profile",
        "description": "Updates the user's profile details after authentication.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Profile"
              }
            }
          }
        }
      },
      "responses": {
        "200": {
          "description": "User profile updated successfully.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Profile"
              }
            }
          }
        },
        "401": {
          "description": "Unauthorized access."
        }
      }
    },
...
  }
}